RavenDB-25324: Introduce DocumentEntryId
Following the work in RavenDB-25323, we're introducing DocumentEntryId to prevent confusion between document entry IDs (index layer) and container entry IDs (storage layer). Both were just longs before, which meant the compiler couldn't catch us if we accidentally mixed them up. To make matters worse, frequency is also encoded into the posting list longs making the errors even more subtle.

This matters because mixing these IDs causes nasty bugs that only show up later: deduplication breaks, documents get indexed twice, the index gets corrupted, or queries return wrong results. The kind of bugs that surface weeks or months after they're introduced.

Now the type system catches these mistakes at compile time. You have to explicitly cast at layer boundaries, which makes it obvious in code review when we're crossing from one layer to another.

Key changes:
- Added DocumentEntryId struct (zero-cost wrapper around long)
- Updated collections to use DocumentEntryId instead of raw longs
- Added typed Encode() overload for document layer operations
- Added comments explaining the encoding duality at critical points

The remaining casts are intentional, mostly at storage boundaries when calling Voron APIs that expect longs.

+++ src/Corax/Indexing/IndexWriter.Testing.cs
@@ -37,7 +37,7 @@ public partial class IndexWriter
                     keysPtr[i] = keys[i];
                     if ((keys[i] & (long)TermIdMask.EnsureIsSingleMask) != 0)
                     {
-                        keysPtr[i] = EntryIdEncodings.GetContainerId(keys[i]);
+                        keysPtr[i] = (long)EntryIdEncodings.GetContainerId(keys[i]);
                         continue;
                     }
