+++ src/Corax/Querying/Matches/TermProviders/TermProvider.TermRange.cs
@@ -251,12 +251,12 @@ public struct TermRangeProvider<TLookupIterator, TLow, THigh> : ITermProvider, I
             
             if ((termId & (long)TermIdMask.PostingList) != 0)
             {
-                postingLists.Add(allocator, EntryIdEncodings.GetContainerId(termId));
+                postingLists.Add(allocator, (long)EntryIdEncodings.GetContainerId(termId));
                 postingListsType.Add(allocator, TermIdMask.PostingList);
             }
             else if ((termId & (long)TermIdMask.SmallPostingList) != 0)
             {
-                postingLists.Add(allocator, EntryIdEncodings.GetContainerId(termId));
+                postingLists.Add(allocator, (long)EntryIdEncodings.GetContainerId(termId));
                 postingListsType.Add(allocator, TermIdMask.SmallPostingList);
             }
             else
